+++ CHANGELOG.md
@@ -10,4 +10,7 @@
 
 ##### Bug Fixes
 
-* None.
+* Detect best migration path correctly when first step of two alternates is
+  an explicit migration.
+  [John Fairhurst](https://github.com/johnfairh)
+  [#5](https://github.com/johnfairh/TMLPersistentContainer/issues/5)